wordCompleterをwordSuggesterに統合
